Codec.Picture.HDR

Description

Module dedicated of Radiance file decompression (.hdr or .pic) file. Radiance file format is used for High dynamic range imaging.

Documentation

decodeHDR :: ByteString -> Either String DynamicImage

Decode an HDR (radiance) image, the resulting image can be:

ImageRGBF

decodeHDRWithMetadata :: ByteString -> Either String (DynamicImage, Metadatas)

Equivalent to decodeHDR but with aditional metadatas.

encodeHDR :: Image PixelRGBF -> ByteString

Encode an High dynamic range image into a radiance image file format. Alias for encodeRawHDR

encodeRawHDR :: Image PixelRGBF -> ByteString

Encode an High dynamic range image into a radiance image file format. without compression

encodeRLENewStyleHDR :: Image PixelRGBF -> ByteString

Encode an High dynamic range image into a radiance image file format using a light RLE compression. Some problems seem to arise with some image viewer.

writeHDR :: FilePath -> Image PixelRGBF -> IO ()

Write an High dynamic range image into a radiance image file on disk.

writeRLENewStyleHDR :: FilePath -> Image PixelRGBF -> IO ()

Write a RLE encoded High dynamic range image into a radiance image file on disk.