Codec.Picture.Jpg.Internal.Progressive

Documentation

data JpgUnpackerParameter

Constructors

JpgUnpackerParameter

Fields

dcHuffmanTree :: !HuffmanPackedTree

acHuffmanTree :: !HuffmanPackedTree

componentIndex :: !Int

restartInterval :: !Int

componentWidth :: !Int

componentHeight :: !Int

subSampling :: !(Int, Int)

coefficientRange :: !(Int, Int)

successiveApprox :: !(Int, Int)

readerIndex :: !Int

indiceVector :: !Int

When in progressive mode, we can have many color in a scan or only one. The indices changes on this fact, when mixed, there is whole MCU for all color components, spanning multiple block lines. With only one color component we use the normal raster order.

blockIndex :: !Int

blockMcuX :: !Int

blockMcuY :: !Int

progressiveUnpack :: (Int, Int) -> JpgFrameHeader -> Vector (MacroBlock Int16) -> [([(JpgUnpackerParameter, a)], ByteString)] -> ST s (MutableImage s PixelYCbCr8)