Codec.Picture.Jpg.Internal.Types

Description

A good explanation of the JPEG format, including diagrams, is given at: https://github.com/corkami/formats/blob/master/image/jpeg.md

The full spec (excluding EXIF): https://www.w3.org/Graphics/JPEG/itu-t81.pdf

Documentation

type MutableMacroBlock s a = STVector s a

Macroblock that can be transformed.

createEmptyMutableMacroBlock :: (Storable a, Num a) => ST s (MutableMacroBlock s a)

Create a new macroblock with the good array size

printMacroBlock :: (Storable a, PrintfArg a) => MutableMacroBlock s a -> ST s String

printPureMacroBlock :: (Storable a, PrintfArg a) => MacroBlock a -> String

type DcCoefficient = Int16

Type only used to make clear what kind of integer we are carrying Might be transformed into newtype in the future

data JpgImage

Constructors

JpgImage

Fields

jpgFrame :: [JpgFrame]

data JpgComponent

Constructors

JpgComponent

Fields

componentIdentifier :: !Word8

horizontalSamplingFactor :: !Word8

Stored with 4 bits

verticalSamplingFactor :: !Word8

Stored with 4 bits

quantizationTableDest :: !Word8

data JpgFrameHeader

Constructors

JpgFrameHeader

Fields

jpgFrameHeaderLength :: !Word16

jpgSamplePrecision :: !Word8

jpgHeight :: !Word16

jpgWidth :: !Word16

jpgImageComponentCount :: !Word8

jpgComponents :: ![JpgComponent]

data JpgFrame

Constructors

JpgAppFrame !Word8 ByteString

JpgAdobeAPP14 !JpgAdobeApp14

JpgJFIF !JpgJFIFApp0

JpgExif ![ImageFileDirectory]

JpgExtension !Word8 ByteString

JpgQuantTable ![JpgQuantTableSpec]

JpgHuffmanTable ![(JpgHuffmanTableSpec, HuffmanPackedTree)]

JpgScanBlob !JpgScanHeader !ByteString
The ByteString is the ECS (Entropy-Coded Segment), typically the largest part of compressed image data.

JpgScans !JpgFrameKind !JpgFrameHeader

JpgIntervalRestart !Word16

data JpgFrameKind

Constructors

JpgBaselineDCTHuffman

JpgExtendedSequentialDCTHuffman

JpgProgressiveDCTHuffman

JpgLosslessHuffman

JpgDifferentialSequentialDCTHuffman

JpgDifferentialProgressiveDCTHuffman

JpgDifferentialLosslessHuffman

JpgExtendedSequentialArithmetic

JpgProgressiveDCTArithmetic

JpgLosslessArithmetic

JpgDifferentialSequentialDCTArithmetic

JpgDifferentialProgressiveDCTArithmetic

JpgDifferentialLosslessArithmetic

JpgQuantizationTable

JpgHuffmanTableMarker

JpgStartOfScan

JpgEndOfImage

JpgAppSegment Word8

JpgExtensionSegment Word8

JpgRestartInterval

JpgRestartIntervalEnd Word8

data JpgScanHeader

Constructors

JpgScanHeader

Fields

scanLength :: !Word16

scanComponentCount :: !Word8

scans :: [JpgScanSpecification]

spectralSelection :: (Word8, Word8)

(begin, end)

successiveApproxHigh :: !Word8

Encoded as 4 bits

successiveApproxLow :: !Word8

Encoded as 4 bits

data JpgQuantTableSpec

Constructors

JpgQuantTableSpec

Fields

quantPrecision :: !Word8

Stored on 4 bits

quantDestination :: !Word8

Stored on 4 bits

quantTable :: MacroBlock Int16

data JpgHuffmanTableSpec

Constructors

JpgHuffmanTableSpec

Fields

huffmanTableClass :: !DctComponent

0 : DC, 1 : AC, stored on 4 bits

huffmanTableDest :: !Word8

Stored on 4 bits

huffSizes :: !(Vector Word8)

huffCodes :: !(Vector (Vector Word8))

data JpgImageKind

Constructors

BaseLineDCT

ProgressiveDCT

data JpgScanSpecification

Constructors

JpgScanSpecification

Fields

componentSelector :: !Word8

dcEntropyCodingTable :: !Word8

Encoded as 4 bits

acEntropyCodingTable :: !Word8

Encoded as 4 bits

data JpgColorSpace

Constructors

JpgColorSpaceYCbCr

JpgColorSpaceYCC

JpgColorSpaceY

JpgColorSpaceYA

JpgColorSpaceYCCA

JpgColorSpaceYCCK

JpgColorSpaceCMYK

JpgColorSpaceRGB

JpgColorSpaceRGBA

data AdobeTransform

Constructors

AdobeUnknown
Value 0

AdobeYCbCr
value 1

AdobeYCck
value 2

data JpgAdobeApp14

Constructors

JpgAdobeApp14

Fields

_adobeDctVersion :: !Word16

_adobeFlag0 :: !Word16

_adobeFlag1 :: !Word16

_adobeTransform :: !AdobeTransform

data JpgJFIFApp0

Constructors

JpgJFIFApp0

Fields

_jfifUnit :: !JFifUnit

_jfifDpiX :: !Word16

_jfifDpiY :: !Word16

_jfifThumbnail :: !(Maybe Int)

data JFifUnit

Size: 1

Constructors

JFifUnitUnknown
0

JFifPixelsPerInch
1

JFifPixelsPerCentimeter
2

newtype TableList a

Type introduced only to avoid some typeclass overlapping problem

Constructors

TableList [a]

newtype RestartInterval

Constructors

RestartInterval Word16

getJpgImage :: Get JpgImage

Like get from `instance Binary JpgImage`, but without the legacy semi-lazy implementation.

calculateSize :: SizeCalculable a => a -> Int

dctBlockSize :: Num a => a

parseECS :: Get ByteString

Parses a Scan's ECS (Entropy-Coded Segment, the largest part of compressed image data) from the Get stream.

When this function is called, the parser's offset should be immediately behind the SOS tag.

As described on e.g. https://www.ccoderun.ca/programming/2017-01-31_jpeg/,

To find the next segment after the SOS, you must keep reading until you
find a 0xFF bytes which is not immediately followed by 0x00 (see "byte stuffing")
[or a reset marker's byte: 0xD0 through 0xD7].
Normally, this will be the EOI segment that comes at the end of the file.

where the 0xFF is the next segment's marker. See https://github.com/corkami/formats/blob/master/image/jpeg.md#entropy-coded-segment for more details.

This function returns the ECS, not including the next segment's marker on its trailing end.

parseECS_simple :: Get ByteString

Simpler implementation of parseECS to allow an easier understanding of the logic, and to provide a comparison for correctness.

skipUntilFrames :: Get ()

skipFrameMarker :: Get ()

parseFrameOfKind :: JpgFrameKind -> Get (Maybe JpgFrame)

Parses a single frame.

Returns Nothing when we encounter a frame we want to skip.

parseFrames :: Get [JpgFrame]

Parse a list of JpgFrames.

parseFrameKinds :: Get [(JpgFrameKind, Int64, Int64)]

Parse a list of JpgFrameKinds with their corresponding offsets and lengths (not counting the segment and frame markers into the lengths).

Useful for debugging.

parseToFirstFrameHeader :: Get (Maybe JpgFrameHeader)

Parses forward, returning the first scan header encountered.

Should be used after skipUntilFrames.

Fails parsing when an SOS segment marker (JpgStartOfScan, resulting in JpgScanBlob) is encountered before an SOF segment marker (that results in JpgScans carrying the JpgFrameHeader).