Codec.Picture.Png.Internal.Type

Description

Low level png module, you should import Internal instead.

Documentation

data PngIHdr

Generic header used in PNG images.

Constructors

PngIHdr

Fields

width :: !Word32

Image width in number of pixel

height :: !Word32

Image height in number of pixel

bitDepth :: !Word8

Number of bit per sample

colourType :: !PngImageType

Kind of png image (greyscale, true color, indexed...)

compressionMethod :: !Word8

Compression method used

filterMethod :: !Word8

Must be 0

interlaceMethod :: !PngInterlaceMethod

If the image is interlaced (for progressive rendering)

data PngFilter

The pixels value should be : +---+---+ | c | b | +---+---+ | a | x | +---+---+ x being the current filtered pixel

Constructors

FilterNone
Filt(x) = Orig(x), Recon(x) = Filt(x)

FilterSub
Filt(x) = Orig(x) - Orig(a), Recon(x) = Filt(x) + Recon(a)

FilterUp
Filt(x) = Orig(x) - Orig(b), Recon(x) = Filt(x) + Recon(b)

FilterAverage
Filt(x) = Orig(x) - floor((Orig(a) + Orig(b)) / 2), Recon(x) = Filt(x) + floor((Recon(a) + Recon(b)) / 2)

FilterPaeth
Filt(x) = Orig(x) - PaethPredictor(Orig(a), Orig(b), Orig(c)), Recon(x) = Filt(x) + PaethPredictor(Recon(a), Recon(b), Recon(c))

data PngInterlaceMethod

Different known interlace methods for PNG image

Constructors

PngNoInterlace
No interlacing, basic data ordering, line by line from left to right.

PngInterlaceAdam7
Use the Adam7 ordering, see adam7Reordering

type PngPalette = Palette' PixelRGB8

Palette with indices beginning at 0 to elemcount - 1

data PngImageType

What kind of information is encoded in the IDAT section of the PngFile

Constructors

PngGreyscale

PngTrueColour

PngIndexedColor

PngGreyscaleWithAlpha

PngTrueColourWithAlpha

data PngPhysicalDimension

Constructors

PngPhysicalDimension

Fields

pngDpiX :: !Word32

pngDpiY :: !Word32

pngUnit :: !PngUnit

newtype PngGamma

Constructors

PngGamma

Fields

getPngGamma :: Double

data PngUnit

Constructors

PngUnitUnknown
0 value

PngUnitMeter
1 value

data APngAnimationControl

Constructors

APngAnimationControl

Fields

animationFrameCount :: !Word32

animationPlayCount :: !Word32

data APngFrameDisposal

Encoded in a Word8

Constructors

APngDisposeNone
No disposal is done on this frame before rendering the next; the contents of the output buffer are left as is. Has Value 0

APngDisposeBackground
The frame's region of the output buffer is to be cleared to fully transparent black before rendering the next frame. Has Value 1

APngDisposePrevious
the frame's region of the output buffer is to be reverted to the previous contents before rendering the next frame. Has Value 2

data APngBlendOp

Encoded in a Word8

Constructors

APngBlendSource
Overwrite output buffer. has value '0'

APngBlendOver
Alpha blend to the output buffer. Has value '1'

data APngFrameControl

Constructors

APngFrameControl

Fields

frameSequenceNum :: !Word32

Starting from 0

frameWidth :: !Word32

Width of the following frame

frameHeight :: !Word32

Height of the following frame

frameLeft :: !Word32

frameTop :: !Word32

frameDelayNumerator :: !Word16

frameDelayDenuminator :: !Word16

frameDisposal :: !APngFrameDisposal

frameBlending :: !APngBlendOp

parsePalette :: PngRawChunk -> Either String PngPalette

Parse a palette from a png chunk.

pngComputeCrc :: [ByteString] -> Word32

Compute the CRC of a raw buffer, as described in annex D of the PNG specification.

pLTESignature :: ChunkSignature

Signature for a palette chunk in the pgn file. Must occure before iDAT.

iDATSignature :: ChunkSignature

Signature for a data chuck (with image parts in it)

iENDSignature :: ChunkSignature

Signature for the last chunk of a png image, telling the end.

tRNSSignature :: ChunkSignature

tEXtSignature :: ChunkSignature

zTXtSignature :: ChunkSignature

gammaSignature :: ChunkSignature

pHYsSignature :: ChunkSignature

animationControlSignature :: ChunkSignature

Low level types

type ChunkSignature = ByteString

Value used to identify a png chunk, must be 4 bytes long.

data PngRawImage

Raw parsed image which need to be decoded.

Constructors

PngRawImage

Fields

header :: PngIHdr

chunks :: [PngRawChunk]

data PngChunk

PNG chunk representing some extra information found in the parsed file.

Constructors

PngChunk

Fields

pngChunkData :: ByteString

The raw data inside the chunk

pngChunkSignature :: ChunkSignature

The name of the chunk.

data PngRawChunk

Data structure during real png loading/parsing

Constructors

PngRawChunk

Fields

chunkLength :: Word32

chunkType :: ChunkSignature

chunkCRC :: Word32

chunkData :: ByteString

data PngLowLevel a

Low level access to PNG information

Constructors

PngLowLevel

Fields

pngImage :: Image a

The real uncompressed image

pngChunks :: [PngChunk]

List of raw chunk where some user data might be present.

chunksWithSig :: PngRawImage -> ChunkSignature -> [ByteString]

mkRawChunk :: ChunkSignature -> ByteString -> PngRawChunk