Codec.Picture.Tiff

Description

Module implementing TIFF decoding.

Supported compression schemes:

Uncompressed

PackBits

LZW

Supported bit depth:

2 bits

4 bits

8 bits

16 bits

Documentation

decodeTiff :: ByteString -> Either String DynamicImage

Decode a tiff encoded image while preserving the underlying pixel type (except for Y32 which is truncated to 16 bits).

This function can output the following images:

ImageY8

ImageY16

ImageY32

ImageYF

ImageYA8

ImageYA16

ImageRGB8

ImageRGB16

ImageRGBA8

ImageRGBA16

ImageCMYK8

ImageCMYK16

decodeTiffWithMetadata :: ByteString -> Either String (DynamicImage, Metadatas)

Like decodeTiff but also provides some metdata present in the Tiff file.

The metadata extracted are the DpiX & DpiY information alongside the EXIF informations.

decodeTiffWithPaletteAndMetadata :: ByteString -> Either String (PalettedImage, Metadatas)

Decode TIFF and provide separated palette and metadata

class Pixel px => TiffSaveable px

Class defining which pixel types can be serialized in a Tiff file.

Minimal complete definition

colorSpaceOfPixel

encodeTiff :: forall px. TiffSaveable px => Image px -> ByteString

Transform an image into a Tiff encoded bytestring, ready to be written as a file.

writeTiff :: TiffSaveable pixel => FilePath -> Image pixel -> IO ()

Helper function to directly write an image as a tiff on disk.