Codec.Picture.Bitmap

Description

Modules used for Bitmap file (.bmp) file loading and writing

Functions

writeBitmap :: BmpEncodable pixel => FilePath -> Image pixel -> IO ()

Write an image in a file use the bitmap format.

encodeBitmap :: forall pixel. BmpEncodable pixel => Image pixel -> ByteString

Encode an image into a bytestring in .bmp format ready to be written on disk.

decodeBitmap :: ByteString -> Either String DynamicImage

Try to decode a bitmap image. Right now this function can output the following pixel types :

PixelRGB8

Pixel8

encodeDynamicBitmap :: DynamicImage -> Either String ByteString

Encode a dynamic image in bmp if possible, supported pixel type are :

RGB8

RGBA8

Y8

writeDynamicBitmap :: FilePath -> DynamicImage -> IO (Either String Bool)

Write a dynamic image in a .bmp image file if possible. The same restriction as encodeDynamicBitmap apply.

Accepted format in output

class BmpEncodable pixel

All the instance of this class can be written as a bitmap file using this library.

Minimal complete definition

bitsPerPixel, bmpEncode