Codec.Picture.Jpg.Internal.Types

Documentation

type MutableMacroBlock s a = STVector s a

Macroblock that can be transformed.

createEmptyMutableMacroBlock :: (Storable a, Num a) => ST s (MutableMacroBlock s a)

Create a new macroblock with the good array size

printMacroBlock :: (Storable a, PrintfArg a) => MutableMacroBlock s a -> ST s String

printPureMacroBlock :: (Storable a, PrintfArg a) => MacroBlock a -> String

type DcCoefficient = Int16

Type only used to make clear what kind of integer we are carrying Might be transformed into newtype in the future

data JpgImage

Constructors

JpgImage

Fields

jpgFrame :: [JpgFrame]

data JpgComponent

Constructors

JpgComponent

Fields

componentIdentifier :: !Word8

horizontalSamplingFactor :: !Word8

Stored with 4 bits

verticalSamplingFactor :: !Word8

Stored with 4 bits

quantizationTableDest :: !Word8

data JpgFrameHeader

Constructors

JpgFrameHeader

Fields

jpgFrameHeaderLength :: !Word16

jpgSamplePrecision :: !Word8

jpgHeight :: !Word16

jpgWidth :: !Word16

jpgImageComponentCount :: !Word8

jpgComponents :: ![JpgComponent]

data JpgFrame

Constructors

JpgAppFrame !Word8 ByteString

JpgAdobeAPP14 !JpgAdobeApp14

JpgJFIF !JpgJFIFApp0

JpgExif ![ImageFileDirectory]

JpgExtension !Word8 ByteString

JpgQuantTable ![JpgQuantTableSpec]

JpgHuffmanTable ![(JpgHuffmanTableSpec, HuffmanPackedTree)]

JpgScanBlob !JpgScanHeader !ByteString

JpgScans !JpgFrameKind !JpgFrameHeader

JpgIntervalRestart !Word16

data JpgFrameKind

Constructors

JpgBaselineDCTHuffman

JpgExtendedSequentialDCTHuffman

JpgProgressiveDCTHuffman

JpgLosslessHuffman

JpgDifferentialSequentialDCTHuffman

JpgDifferentialProgressiveDCTHuffman

JpgDifferentialLosslessHuffman

JpgExtendedSequentialArithmetic

JpgProgressiveDCTArithmetic

JpgLosslessArithmetic

JpgDifferentialSequentialDCTArithmetic

JpgDifferentialProgressiveDCTArithmetic

JpgDifferentialLosslessArithmetic

JpgQuantizationTable

JpgHuffmanTableMarker

JpgStartOfScan

JpgEndOfImage

JpgAppSegment Word8

JpgExtensionSegment Word8

JpgRestartInterval

JpgRestartIntervalEnd Word8

data JpgScanHeader

Constructors

JpgScanHeader

Fields

scanLength :: !Word16

scanComponentCount :: !Word8

scans :: [JpgScanSpecification]

spectralSelection :: (Word8, Word8)

(begin, end)

successiveApproxHigh :: !Word8

Encoded as 4 bits

successiveApproxLow :: !Word8

Encoded as 4 bits

data JpgQuantTableSpec

Constructors

JpgQuantTableSpec

Fields

quantPrecision :: !Word8

Stored on 4 bits

quantDestination :: !Word8

Stored on 4 bits

quantTable :: MacroBlock Int16

data JpgHuffmanTableSpec

Constructors

JpgHuffmanTableSpec

Fields

huffmanTableClass :: !DctComponent

0 : DC, 1 : AC, stored on 4 bits

huffmanTableDest :: !Word8

Stored on 4 bits

huffSizes :: !(Vector Word8)

huffCodes :: !(Vector (Vector Word8))

data JpgImageKind

Constructors

BaseLineDCT

ProgressiveDCT

data JpgScanSpecification

Constructors

JpgScanSpecification

Fields

componentSelector :: !Word8

dcEntropyCodingTable :: !Word8

Encoded as 4 bits

acEntropyCodingTable :: !Word8

Encoded as 4 bits

data JpgColorSpace

Constructors

JpgColorSpaceYCbCr

JpgColorSpaceYCC

JpgColorSpaceY

JpgColorSpaceYA

JpgColorSpaceYCCA

JpgColorSpaceYCCK

JpgColorSpaceCMYK

JpgColorSpaceRGB

JpgColorSpaceRGBA

data AdobeTransform

Constructors

AdobeUnknown
Value 0

AdobeYCbCr
value 1

AdobeYCck
value 2

data JpgAdobeApp14

Constructors

JpgAdobeApp14

Fields

_adobeDctVersion :: !Word16

_adobeFlag0 :: !Word16

_adobeFlag1 :: !Word16

_adobeTransform :: !AdobeTransform

data JpgJFIFApp0

Constructors

JpgJFIFApp0

Fields

_jfifUnit :: !JFifUnit

_jfifDpiX :: !Word16

_jfifDpiY :: !Word16

_jfifThumbnail :: !(Maybe Int)

data JFifUnit

Size: 1

Constructors

JFifUnitUnknown
0

JFifPixelsPerInch
1

JFifPixelsPerCentimeter
2

calculateSize :: SizeCalculable a => a -> Int

dctBlockSize :: Num a => a