Codec.Picture.Tiff

Description

Module implementing TIFF decoding.

Supported compression schemes:

Uncompressed

PackBits

LZW

Supported bit depth:

2 bits

4 bits

8 bits

16 bits

Documentation

decodeTiff :: ByteString -> Either String DynamicImage

Decode a tiff encoded image while preserving the underlying pixel type (except for Y32 which is truncated to 16 bits).

This function can output the following pixel types:

PixelY8

PixelY16

PixelYA8

PixelYA16

PixelRGB8

PixelRGB16

PixelRGBA8

PixelRGBA16

PixelCMYK8

PixelCMYK16

class Pixel px => TiffSaveable px

Class defining which pixel types can be serialized in a Tiff file.

Minimal complete definition

colorSpaceOfPixel

encodeTiff :: forall px. TiffSaveable px => Image px -> ByteString

Transform an image into a Tiff encoded bytestring, ready to be written as a file.

writeTiff :: TiffSaveable pixel => FilePath -> Image pixel -> IO ()

Helper function to directly write an image as a tiff on disk.