Codec.Picture.Tiff.Internal.Types

Documentation

class BinaryParam a b where

Because having a polymorphic get with endianness is to nice to pass on, introducing this helper type class, which is just a superset of Binary, but formalising a parameter passing into it.

Methods

getP :: a -> Get b

putP :: a -> b -> Put

data Endianness

Constructors

EndianLittle

EndianBig

data TiffHeader

Constructors

TiffHeader

Fields

hdrEndianness :: !Endianness

hdrOffset :: !Word32

data TiffPlanarConfiguration

Constructors

PlanarConfigContig

PlanarConfigSeparate

data TiffCompression

Constructors

CompressionNone

CompressionModifiedRLE

CompressionLZW

CompressionJPEG

CompressionPackBit

data IfdType

Constructors

TypeByte

TypeAscii

TypeShort

TypeLong

TypeRational

TypeSByte

TypeUndefined

TypeSignedShort

TypeSignedLong

TypeSignedRational

TypeFloat

TypeDouble

data TiffColorspace

Constructors

TiffMonochromeWhite0
0

TiffMonochrome
1

TiffRGB
2

TiffPaleted
3

TiffTransparencyMask
4

TiffCMYK
5

TiffYCbCr
6

TiffCIELab
8

data TiffSampleFormat

Constructors

TiffSampleUint

TiffSampleInt

TiffSampleFloat

TiffSampleUnknown

data ImageFileDirectory

Constructors

ImageFileDirectory

Fields

ifdIdentifier :: !ExifTag

ifdType :: !IfdType

ifdCount :: !Word32

ifdOffset :: !Word32

ifdExtended :: !ExifData

data ExtraSample

Constructors

ExtraSampleUnspecified
0

ExtraSampleAssociatedAlpha
1

ExtraSampleUnassociatedAlpha
2

data Predictor

Constructors

PredictorNone

PredictorHorizontalDifferencing

planarConfgOfConstant :: Word32 -> Get TiffPlanarConfiguration

constantToPlaneConfiguration :: TiffPlanarConfiguration -> Word16

unpackSampleFormat :: Word32 -> Get TiffSampleFormat

packSampleFormat :: TiffSampleFormat -> Word32

word16OfTag :: ExifTag -> Word16

Convert a tag to it's corresponding value.

unpackPhotometricInterpretation :: Word32 -> Get TiffColorspace

packPhotometricInterpretation :: TiffColorspace -> Word16

codeOfExtraSample :: ExtraSample -> Word16

unPackCompression :: Word32 -> Get TiffCompression

packCompression :: TiffCompression -> Word16

predictorOfConstant :: Word32 -> Get Predictor